Split Group typeclass to separate module
